+++ src/parameterized_tendencies/gravity_wave_drag/orographic_gravity_wave_helper.jl
@@ -226,10 +226,13 @@ end
 
 function compute_OGW_info(Y, elev_data, earth_radius, γ, h_frac)
     # obtain lat, lon, elevation from the elev_data
+    FT = Spaces.undertype(Spaces.axes(Y.c))
+    # downsample to elev dims (3600×1800)
+    skip_pt = 6
     nt = NCDataset(elev_data, "r") do ds
-        lon = Array(ds["longitude"])
-        lat = Array(ds["latitude"])
-        elev = Array(ds["elevation"])
+        lon = FT.(Array(ds["lon"]))[1:skip_pt:end]
+        lat = FT.(Array(ds["lat"]))[1:skip_pt:end]
+        elev = FT.(Array(ds["z"]))[1:skip_pt:end, 1:skip_pt:end]
         (; lon, lat, elev)
     end
     (; lon, lat, elev) = nt