+++ test/restart.jl
@@ -69,6 +69,11 @@ function _error(
     arr2::AbstractArray;
     ABS_TOL = 100eps(eltype(arr1)),
 )
+    # There are some parameters, e.g. Obukhov length, for which Inf
+    # is a reasonable value (implying a stability parameter in the neutral boundary layer
+    # regime, for instance). We account for such instances with the `isfinite` function.
+    arr1 .*= isfinite.(arr1)
+    arr2 .*= isfinite.(arr2)
     diff = abs.(arr1 .- arr2)
     denominator = abs.(arr1)
     error = ifelse.(denominator .> ABS_TOL, diff ./ denominator, diff)